refactor: LottoGame 클래스를 domain 패키지로 이동
